Moved Pages all together and created new projects for Game and MainMenu
