+++ MauiGeoQuiz.Game/Platforms/Tizen/PlatformClass1.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace MauiGeoQuiz.Game;
+// All the code in this file is only included on Tizen.
+public class PlatformClass1
+{
+}
