+++ MauiGeoQuiz.MainMenu/Platforms/Android/PlatformClass1.cs
@@ -1,0 +1,6 @@
+﻿namespace MauiGeoQuiz.MainMenu;
+
+// All the code in this file is only included on Android.
+public class PlatformClass1
+{
+}
