+++ MauiGeoQuiz.MainMenu/Platforms/MacCatalyst/PlatformClass1.cs
@@ -1,0 +1,6 @@
+﻿namespace MauiGeoQuiz.MainMenu;
+
+// All the code in this file is only included on Mac Catalyst.
+public class PlatformClass1
+{
+}
