+++ MauiGeoQuiz.MainMenu/Platforms/Tizen/PlatformClass1.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace MauiGeoQuiz.MainMenu;
+// All the code in this file is only included on Tizen.
+public class PlatformClass1
+{
+}
