+++ MauiGeoQuiz.MainMenu/Platforms/Windows/PlatformClass1.cs
@@ -1,0 +1,6 @@
+﻿namespace MauiGeoQuiz.MainMenu;
+
+// All the code in this file is only included on Windows.
+public class PlatformClass1
+{
+}
