+++ MauiGeoQuiz/ViewPages/GamePage.xaml
@@ -7,7 +7,7 @@
     xmlns:strings="clr-namespace:MauiGeoQuiz.Resources.Strings"
     x:TypeArguments="vm:GameViewModel"
     x:DataType="vm:GameViewModel"
-    x:Class="MauiGeoQuiz.Game.GamePage">
+    x:Class="MauiGeoQuiz.ViewPages.GamePage">
 
     <ScrollView>
